pyproject.toml
[tool.poetry]
name = "zoomcamp"
version = "0.1.0"
description = ""
authors = ["Jithish J <135303624+jithsg@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.0"
scikit-learn = "^1.4.0"
pyarrow = ">=4.0.0,<15"
mlflow = "^2.9.2"
xgboost = "^2.0.3"
flask = "^3.0.1"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
requests = "^2.31.0"
hyperopt = "^0.2.7"
boto3 = "^1.34.27"
prefect = "^2.14.17"
prefect-aws = "^0.4.9"
jupyter = "^1.0.0"
seaborn = "^0.13.1"
orjson = "^3.9.12"
httpx = "^0.26.0"
black = "^24.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
